LayoutTests/fast/css-custom-paint/registerPaintBindings.html

<!DOCTYPE html><!-- webkit-test-runner [ experimental:CSSPaintingAPIEnabled=true ] -->
<meta name="author" title="Justin Michaud" href="mailto:justin_michaud@webkit.org">
<meta name="assert" content="registerPaint accesses arguments in the correct order">
<link rel="help" content="https://drafts.css-houdini.org/css-paint-api-1/">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
class MyPaint {
  paint() { }
}

test(function () {
    assert_true('registerPaint' in CSS.paintWorkletGlobalScope.__proto__, '"registerPaint" exists on prototype');
    assert_true('registerPaint' in CSS.paintWorkletGlobalScope, '"registerPaint" exists on window');
}, 'registerPaint must be a method');

test(function () {
    const calls = [];
    const proxy = new Proxy(class extends MyPaint { }, {
        get: function (target, name) {
            calls.push(name);
            return target[name];
        }
    });
    CSS.paintWorkletGlobalScope.registerPaint('test-proto', proxy);
    assert_array_equals(calls, ["inputProperties", "inputArguments", "contextOptions", "prototype"]);
}, 'must get "prototype" property of the constructor');

test(function () {
    const calls = [];
    const proxy = new Proxy(class extends MyPaint { }, {
        get: function (target, name) {
            calls.push(name);
            if (name == 'inputProperties')
                throw {name: 'expectedError'};
            return target[name];
        }
    });
    assert_throws({'name': 'expectedError'}, function () { CSS.paintWorkletGlobalScope.registerPaint('test-rethrow-inputProperties0', proxy); });
}, 'must rethrow an exception thrown while getting "inputProperties" property of the constructor');

test(function () {
    const calls = [];
    const proxy = new Proxy(class extends MyPaint { }, {
        get: function (target, name) {
            calls.push(name);
            if (name == 'prototype')
                throw {name: 'expectedError'};
            return target[name];
        }
    });
    assert_throws({'name': 'expectedError'}, function () { CSS.paintWorkletGlobalScope.registerPaint('test-rethrow-proto', proxy); });
}, 'must rethrow an exception thrown while getting "prototype" property of the constructor');

test(function () {
    let returnedValue;
    const proxy = new Proxy(class extends MyPaint { }, {
        get: function (target, name) {
          if (name == 'prototype')
            return returnedValue;
          return target[name];
        }
    });

    returnedValue = null;
    assert_throws({'name': 'TypeError'}, function () { CSS.paintWorkletGlobalScope.registerPaint('test-rethrow-proto-noobj1', proxy); },
        'must throw when "prototype" property of the constructor is null');
    returnedValue = undefined;
    assert_throws({'name': 'TypeError'}, function () { CSS.paintWorkletGlobalScope.registerPaint('test-rethrow-proto-noobj2', proxy); },
        'must throw when "prototype" property of the constructor is undefined');
    returnedValue = 'hello';
    assert_throws({'name': 'TypeError'}, function () { CSS.paintWorkletGlobalScope.registerPaint('test-rethrow-proto-noobj3', proxy);; },
        'must throw when "prototype" property of the constructor is a string');
    returnedValue = 1;
    assert_throws({'name': 'TypeError'}, function () { CSS.paintWorkletGlobalScope.registerPaint('test-rethrow-proto-noobj4', proxy);; },
        'must throw when "prototype" property of the constructor is a number');

}, 'must throw when "prototype" property of the constructor is not an object');

test(function () {
    const constructor = function () {}
    constructor.prototype.paint = function() {}
    const calls = [];
    constructor.prototype = new Proxy(constructor.prototype, {
        get: function (target, name) {
            calls.push(name);
            return target[name];
        }
    });
    CSS.paintWorkletGlobalScope.registerPaint('callbacks', constructor);
    assert_array_equals(calls, ['paint']);
}, 'must get paint callback of the constructor prototype');

test(function () {
    const constructor = function () {}
    const calls = [];
    constructor.prototype = new Proxy(constructor.prototype, {
        get: function (target, name) {
            calls.push(name);
            if (name == 'paint')
                throw {name: 'expectedError'};
            return target[name];
        }
    });
    assert_throws({'name': 'expectedError'}, function () { CSS.paintWorkletGlobalScope.registerPaint('callbacks-throw', constructor); });
    assert_array_equals(calls, ['paint']);
}, 'must rethrow an exception thrown while getting paint callback on the constructor prototype');

test(function () {
    const constructor = function () {}
    const calls = [];
    constructor.prototype = new Proxy(constructor.prototype, {
        get: function (target, name) {
            calls.push(name);
            if (name == 'paint')
                return 1;
            return target[name];
        }
    });
    assert_throws({'name': 'TypeError'}, function () { CSS.paintWorkletGlobalScope.registerPaint('callbacks-throw2', constructor); });
    assert_array_equals(calls, ['paint']);
}, 'must rethrow an exception thrown while converting paint callback value to Function callback type');

test(function () {
  const calls = [];
  const proxy = new Proxy(class extends MyPaint { }, {
      get: function (target, name) {
          calls.push(name);
          if (name == 'inputArguments')
            return 1;
          return target[name];
      }
  });
  assert_throws({'name': 'TypeError'}, function () { CSS.paintWorkletGlobalScope.registerPaint('sequence-throw', proxy); });
  assert_array_equals(calls, ["inputProperties", "inputArguments"]);
}, 'must rethrow an exception thrown while converting the value of inputArguments to sequence<DOMString>');

test(function () {
    const constructor = function () {}
    constructor.inputArguments = {[Symbol.iterator]: function *() {
        yield '<length>';
        throw {name: 'SomeError'};
    }};
    assert_throws({'name': 'SomeError'}, function () { CSS.paintWorkletGlobalScope.registerPaint('sequence-throw2', constructor); });
}, 'must rethrow an exception thrown while iterating over inputArguments to sequence<DOMString>');

test(function () {
    const constructor = function () {}
    constructor.inputArguments = {[Symbol.iterator]: 1};
    assert_throws({'name': 'TypeError'}, function () { CSS.paintWorkletGlobalScope.registerPaint('symbol-iterator-throw2', constructor); });
}, 'must rethrow an exception thrown while retrieving Symbol.iterator on inputArguments');
</script>
